move accessible color function to plugins
